pyproject.toml
[build-system]
requires = [
    "setuptools >= 61.2",
    "setuptools_scm[toml] >= 6.2"
]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true
packages = ["umi"]

[project]
name = "umi"
authors = [{name = "Zero ASIC"}]
description = "Universal Memory Interface (UMI)"
readme = "README.md"
urls = {Homepage = "https://github.com/zeroasiccorp/umi"}
requires-python = ">= 3.9"
license = {file = "LICENSE"}
dependencies = [
    "siliconcompiler >= 0.36.5",
    "lambdalib >= 0.4.0, < 0.11.0"
]
dynamic = ["version"]

[tool.setuptools.dynamic]
version = {attr = "umi.__version__"}

[project.optional-dependencies]
test = [
    "pytest == 8.4.2",
    "pytest-xdist == 3.8.0",
    "pytest-timeout == 2.4.0",
    "flake8 == 7.3.0",
    "switchboard-hw == 0.3.0",
    "cocotb==2.0.1",
    "cocotb-bus==0.3.0",
    "cocotbext-axi==0.1.26"
]

[tool.check-wheel-contents]
ignore = [
    "W002"
]

[tool.pytest.ini_options]
markers = [
    "switchboard: this test requires switchboard to run"
]
testpaths = [
    "tests",
    "umi/sumi/tests",
    "umi/lumi/tests"
]
timeout = "360"
